+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Logging improvement.
+### Upgraded
+- Keyple Distributed Network Lib `2.3.1` -> `2.3.2` (source code not impacted)
 
 ## [2.3.1] - 2024-04-12
 ### Changed